+++ src/common/profile.ui.ts
@@ -49,6 +49,8 @@ let _profileLoadExpand = true;
 let _profileLoadSelect = 'default';
 let _iComments = true;
 let _bComments = true;
+let _iCommentsStr = ['/', '/'];
+let _bCommentsStr = ['/', '*'];
 
 
 const _controllers = {};
@@ -2703,8 +2705,14 @@ function loadOptions() {
     _stacking = options.commandStackingChar;
     _speed = options.speedpathsChar;
     _verbatim = options.verbatimChar;
-    _iComments = options.allowInlineComments;
-    _bComments = options.allowBlockComments;
+    _iComments = options.enableInlineComments;
+    _bComments = options.enableBlockComments;
+    _iCommentsStr = options.inlineCommentString.split('');
+    _bCommentsStr = options.blockCommentString.split('');
+    if (_iCommentsStr.length === 1)
+        _iCommentsStr.push(_iCommentsStr[0]);
+    if (_bCommentsStr.length === 1)
+        _bCommentsStr.push(_bCommentsStr[0]);
     _profileLoadExpand = options.profiles.profileExpandSelected;
     _profileLoadSelect = options.profiles.profileSelected;
     if (!profiles.contains(_profileLoadSelect))
@@ -4003,13 +4011,18 @@ function setParseSyntax(editor) {
         //console.log(rules);
         if (Object.prototype.hasOwnProperty.call(rules, 'start')) {
             var b = rules['start'].pop();
-            if(!_iComments)
-            {
+            if (!_iComments) {
                 rules['start'].pop();
                 rules['start'].pop();
-            }   
-            if(_bComments)
+            }
+            else {
+                rules['start'][rules['start'].length - 2].regex = `\\${_iCommentsStr[0]}\\${_iCommentsStr[1]}$`;
+                rules['start'][rules['start'].length - 1].regex = `\\${_iCommentsStr[0]}\\${_iCommentsStr[1]}`;
+            }
+            if (_bComments) {
+                b.regex = `\\${_bCommentsStr[0]}\\${_bCommentsStr[1]}`;
                 rules['start'].push(b);
+            }
             rules['start'][3].token = _stacking;
             rules['start'][3].regex = _stacking;
             rules['start'][5].regex = _parameter + rules['start'][5].regex.substr(1);
@@ -4084,6 +4097,9 @@ function setParseSyntax(editor) {
 7: {token: 'constant.numeric', regex: '[+-]?\\d+(?:(?:\\.\\d*)?(?:[eE][+-]?\\d+)?)?\\b', onMatch: null}
 8: {token: 'text', regex: '\\s+', next: 'start', onMatch: null}      
             */
+        }
+        if (_bComments && Object.prototype.hasOwnProperty.call(rules, 'comment')) {
+            rules['comment'][0].regex = `\\${_bCommentsStr[1]}\\${_bCommentsStr[0]}`;
         }
         //console.log(rules);
         // force recreation of tokenizer