+++ changelog.md
@@ -7,8 +7,10 @@
   - Add [#NOOP](docs/commands.md#miscellaneous) does nothing, will expand arguments and execute any functions but return nothing to the screen
   - [Preferences](docs/preferences.md):
     - Add [Ignore Eval Undefined](docs/preferences.md#scripting) Will cause undefined results to be ignored if enabled, else it will be converted to a string and the word undefined displayed
-    - Add [Allow Inline Comments](docs/preferences.md#scripting) Allow using // inline style comments, anything following // will be ignored until newline
-    - Add [Allow Block Comments](docs/preferences.md#scripting) Allow using /* \*/ block style comments, anything between /* and \*/ will be ignored    
+    - [Inline Comment String](docs/preferences.md#scripting--special-characters) The 2 character string for inline comments, if only 1 character supplied, it is doubled up
+    - [Enable Inline Comment](docs/preferences.md#scripting--special-characters) disable inline comment support
+    - [Block Comment String](docs/preferences.md#scripting--special-characters) The 2 character string for block comments, if only 1 character supplied, it is doubled up
+    - [Block Inline Comment](docs/preferences.md#scripting--special-characters) disable block comment support    
 - **Fixed:**
   - Added type checks to [#ADD](docs/commands.md#miscellaneous) and [#MATH](docs/commands.md#miscellaneous) to ensure the resulting value is a number
   - Added type check to [#CASE](docs/commands.md#conditionals) and [#REPEAT](docs/commands.md#repeating-and-loops) to ensure the value is a number